requirements.txt
google-genai
python-dotenv
Pillow
PyMuPDF
fastapi
uvicorn
pyjwt
python-multipart
passlib[bcrypt]
pandas
streamlit
requests